have simulator as a component on loading
