+++ front-end/src/components/forms/form-validation/FVLogin.tsx
@@ -77,7 +77,7 @@ const FVRegister = () => {
         <FormGroup>
           <FormControlLabel
             control={<CustomCheckbox defaultChecked />}
-            label="Remeber this Device"
+            label="Remember this Device"
           />
         </FormGroup>
         <Typography
